Add sort by members count and show members count in about (#280)
* Add sort by members and show members count in about chat

* Update type for followersCount

* Show members count in chat navbar if any

* Refactor code

* Move community folder outside of modals

* Move form inside its own component
