+++ src/components/modals/about/AboutChatModal.tsx
@@ -1,4 +1,5 @@
 import { ActionCardProps } from '@/components/ActionCard'
+import UpsertChatModal from '@/components/community/UpsertChatModal'
 import ProfilePreview from '@/components/ProfilePreview'
 import ProfilePreviewModalWrapper from '@/components/ProfilePreviewModalWrapper'
 import TruncatedText from '@/components/TruncatedText'
@@ -26,7 +27,6 @@ import {
 } from 'react-icons/hi2'
 import { RxEnter, RxExit } from 'react-icons/rx'
 import urlJoin from 'url-join'
-import UpsertChatModal from '../community/UpsertChatModal'
 import ConfirmationModal from '../ConfirmationModal'
 import MetadataModal from '../MetadataModal'
 import { ModalFunctionalityProps } from '../Modal'
@@ -179,6 +179,11 @@ export default function AboutChatModal({
 
   const closeModal = () => setOpenedModalType(null)
 
+  let subtitle = `${messageCount} messages`
+  if (chat.struct.followersCount) {
+    subtitle += `, ${chat.struct.followersCount} followers`
+  }
+
   return (
     <>
       <JoinChatWrapper>
@@ -189,7 +194,7 @@ export default function AboutChatModal({
               id={chat.id}
               isOpen={props.isOpen && openedModalType === null}
               title={content?.title}
-              subtitle={`${messageCount} messages`}
+              subtitle={subtitle}
               actionMenu={getActionMenu(mutateAsync, isLoading)}
               contentList={contentList}
               image={content?.image}
@@ -234,10 +239,10 @@ export default function AboutChatModal({
         urlTitle={content.title}
       />
       <UpsertChatModal
-        chat={chat}
         isOpen={openedModalType === 'edit'}
         closeModal={closeModal}
         onBackClick={() => setOpenedModalType(null)}
+        formProps={{ chat }}
       />
       <HideUnhideChatWrapper>
         {({ isLoading, mutateAsync }) => {
