+++ src/pages/[hubId]/[slug]/about.tsx
@@ -25,12 +25,11 @@ export const getStaticProps = getCommonStaticProps<AppCommonProps>(
 
     try {
       const [chat] = await getPostsServer([chatId])
+      const content = chat.content
       if (!chat.struct.hidden) {
-        title = chat.content?.title ?? ''
-        description = chat.content?.body ?? ''
-        image = chat.content?.image
-          ? getIpfsContentUrl(chat.content?.image)
-          : ''
+        title = content?.title ?? ''
+        description = content?.body ?? ''
+        image = content?.image ? getIpfsContentUrl(content?.image) : ''
       }
     } catch (err) {
       console.error('Error fetching for chat about page: ', err)
